Create 2.3 Terminating Fift.md
2.3 Terminating Fift
The wordbyeterminates the Fift interpreter with a zero exit code. If a
non-zero exit code is required (for instance, in Fift scripts), one can use word
halt, which terminates Fift with the given exit code (passed as anInteger
at the top of the stack). In contrast,quitdoes not quit to the operating
system, but rather exits to the top level of the Fift interpreter.